interpolate + caleidoscope demohässäkkään
git-svn-id: svn+ssh://melkinpaasi/fs-1/1/kiirala/repo-svn/demotool@65 15d524df-0ca0-45ad-9872-4ba5b019094f

+++ superformula.h
@@ -23,8 +23,8 @@ public:
   void render(double time);
 
   // a, b, m, n1, n2, n3
-  double r1[6];
-  double r2[6];
+  GLfloat r1[6];
+  GLfloat r2[6];
 
   GLfloat scale;
